Step 7 feat: create Task, controllers, rebuild User and TaskStatus for this

+++ src/main/java/hexlet/code/service/TaskStatusService.java
@@ -6,8 +6,10 @@ import hexlet.code.dto.TaskStatusUpdateDTO;
 import hexlet.code.exception.ResourceNotFoundException;
 import hexlet.code.mapper.TaskStatusMapper;
 import hexlet.code.model.TaskStatus;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.TaskStatusRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -17,6 +19,9 @@ import java.util.List;
 public class TaskStatusService {
     @Autowired
     private TaskStatusRepository taskStatusRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     @Autowired
     private TaskStatusMapper taskStatusMapper;
@@ -51,6 +56,9 @@ public class TaskStatusService {
     }
 
     public void delete(Long id) {
+        if (taskRepository.findByTaskStatusId(id).isPresent()) {
+            throw new DataIntegrityViolationException("Status is used by a task and cannot be deleted.");
+        }
         taskStatusRepository.deleteById(id);
     }
